--- tests/unit/bases/all.ts
@@ -1,1 +1,0 @@
-import './createWidgetBase';
